refactor(rust): extract query + aggregate into db/ (slice 5/9) (#15)
* refactor(rust): extract query + aggregate into db/ (slice 5/9)

Move find-style and aggregation query logic out of the lib.rs god-module into dedicated submodules:

- db/query.rs: execute_mql_query_impl, count_documents_impl, explain_mql_query_impl
- db/aggregate.rs: execute_aggregate_impl, explain_aggregate_query_impl

The thin #[tauri::command] wrappers stay in lib.rs and call the impls via crate-root re-exports (pub use db::query::{...}, db::aggregate::{...}), so command registration and the test suite are unchanged. Pure move; no behavior change. cargo test: 60 passed.

* refactor(rust): extract db metadata operations (slice 6/9)

Move database, collection, and index metadata operations into db/metadata.rs:

- list_databases_impl
- list_collections_impl
- list_indexes_impl
- create_index_impl
- delete_index_impl

Keep crate-root re-exports so Tauri command wrappers and tests continue to call the same symbols. Behavior unchanged.

cargo check: passed
cargo test: 60 passed

* refactor(rust): extract collection export jobs (slice 7/9)

Move collection export task orchestration and CSV/JSON file-writing helpers into db/export.rs:

- start_collection_export_impl
- export_mock_collection_to_file
- export_real_collection_to_file
- task progress helpers
- CSV/JSON conversion helpers

Keep the Tauri command wrapper in lib.rs and re-export the impl from the crate root. Behavior unchanged.

cargo check: passed
cargo test: 60 passed

* refactor(rust): extract document mutations (slice 8/9)

Move document mutation and import operations into db/documents.rs:

- json_to_bson_document
- delete_document_impl, delete_many_impl
- update_many_impl, insert_document_impl, update_document_impl
- import_documents_impl and ImportResult

Keep crate-root re-exports so command wrappers and tests continue to use the same symbols. Behavior unchanged.

cargo check: passed
cargo test: 60 passed

* refactor(rust): extract db ddl operations (slice 9/9)

Move collection, view, and database DDL operations into db/ddl.rs:

- create_collection_impl, create_view_impl
- drop_collection_impl, rename_collection_impl
- drop_database_impl, rename_database_impl
- DatabaseRenameResult

Keep crate-root re-exports so command wrappers and tests continue to use the same symbols. Behavior unchanged.

cargo check: passed
cargo test: 60 passed
