+++ src-tauri/src/db/mod.rs
@@ -1,5 +1,11 @@
 //! Database-operation modules (split out of lib.rs).
 
+pub mod aggregate;
+pub mod ddl;
+pub mod documents;
+pub mod export;
 pub mod gridfs;
+pub mod metadata;
+pub mod query;
 pub mod schema;
 pub mod version;
