+++ src-tauri/src/lib.rs
@@ -1,10 +1,9 @@
 use mongodb::{options::ClientOptions, Client};
 use serde::Serialize;
 use serde_json;
-use std::collections::{BTreeSet, HashMap};
 use std::process::Stdio;
-use std::sync::{Arc, Mutex};
-use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+use std::sync::Arc;
+use std::time::{Duration, Instant};
 use tokio::io::{AsyncBufReadExt, AsyncRead, AsyncWriteExt, BufReader};
 use tokio::process::{Child, ChildStdin, Command as TokioCommand};
 use tokio::sync::{mpsc, Mutex as AsyncMutex};
@@ -12,14 +11,29 @@ use uuid::Uuid;
 
 pub mod ai;
 pub mod connections;
+mod db;
 pub(crate) mod mock_db;
 pub mod queries;
 pub mod ssh_tunnel;
-mod vault;
-mod db;
 mod state;
+mod vault;
 mod window;
+pub use db::aggregate::{execute_aggregate_impl, explain_aggregate_query_impl};
+pub use db::ddl::{
+    create_collection_impl, create_view_impl, drop_collection_impl, drop_database_impl,
+    rename_collection_impl, rename_database_impl, DatabaseRenameResult,
+};
+pub use db::documents::{
+    delete_document_impl, delete_many_impl, import_documents_impl, insert_document_impl,
+    json_to_bson_document, update_document_impl, update_many_impl, ImportResult,
+};
+pub use db::export::start_collection_export_impl;
 pub use db::gridfs::{download_gridfs_file_impl, list_gridfs_files_impl, GridFsFileInfo};
+pub use db::metadata::{
+    create_index_impl, delete_index_impl, list_collections_impl, list_databases_impl,
+    list_indexes_impl,
+};
+pub use db::query::{count_documents_impl, execute_mql_query_impl, explain_mql_query_impl};
 pub use db::schema::{analyze_schema_impl, infer_schema, FieldStat, SchemaReport, TypeCount};
 pub use db::version::get_mongodb_version_impl;
 pub use state::{AppState, LockExt};
@@ -118,40 +132,6 @@ pub struct TaskInfo {
     pub error: Option<String>,
     pub created_at_ms: u64,
     pub finished_at_ms: Option<u64>,
-}
-
-fn now_ms() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_millis() as u64
-}
-
-fn update_task<F>(tasks: &Arc<Mutex<HashMap<String, TaskInfo>>>, task_id: &str, update: F)
-where
-    F: FnOnce(&mut TaskInfo),
-{
-    if let Some(task) = tasks.lock().unwrap_or_else(|p| p.into_inner()).get_mut(task_id) {
-        update(task);
-    }
-}
-
-fn fail_task(tasks: &Arc<Mutex<HashMap<String, TaskInfo>>>, task_id: &str, err: String) {
-    update_task(tasks, task_id, |task| {
-        task.status = "failed".to_string();
-        task.message = "Export failed".to_string();
-        task.error = Some(err);
-        task.finished_at_ms = Some(now_ms());
-    });
-}
-
-fn finish_task(tasks: &Arc<Mutex<HashMap<String, TaskInfo>>>, task_id: &str, processed: u64) {
-    update_task(tasks, task_id, |task| {
-        task.status = "completed".to_string();
-        task.processed = processed;
-        task.message = "Export complete".to_string();
-        task.finished_at_ms = Some(now_ms());
-    });
 }
 
 /// Probe each local agent's binary with `--version` (short, blocking). NotFound -> not available.
@@ -527,1003 +507,6 @@ pub async fn stop_mongosh_session_impl(state: &AppState, session_id: &str) -> Re
     Ok(())
 }
 
-pub async fn list_databases_impl(state: &AppState, id: &str) -> Result<Vec<String>, String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return Ok(vec![
-            "admin".to_string(),
-            "config".to_string(),
-            "local".to_string(),
-            "sales_db".to_string(),
-            "user_analytics".to_string(),
-        ]);
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let dbs = client
-        .list_database_names()
-        .await
-        .map_err(|e| format!("Failed to list databases: {}", e))?;
-
-    Ok(dbs)
-}
-
-pub async fn list_collections_impl(
-    state: &AppState,
-    id: &str,
-    db: &str,
-) -> Result<Vec<CollectionInfo>, String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return Ok(mock_db::get_mock_collections(db)
-            .into_iter()
-            .map(|name| CollectionInfo {
-                name,
-                collection_type: "collection".to_string(),
-            })
-            .collect());
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let database = client.database(db);
-    // Use list_collections (not list_collection_names) so we can read each
-    // collection's type and let the UI separate Collections / Views / etc.
-    let mut cursor = database
-        .list_collections()
-        .await
-        .map_err(|e| format!("Failed to list collections: {}", e))?;
-
-    let mut collections = Vec::new();
-    use futures::stream::StreamExt;
-    while let Some(result) = cursor.next().await {
-        let spec = result.map_err(|e| format!("Collection read error: {}", e))?;
-        let collection_type = match spec.collection_type {
-            mongodb::results::CollectionType::View => "view",
-            mongodb::results::CollectionType::Timeseries => "timeseries",
-            _ => "collection",
-        };
-        collections.push(CollectionInfo {
-            name: spec.name,
-            collection_type: collection_type.to_string(),
-        });
-    }
-
-    Ok(collections)
-}
-
-pub async fn list_indexes_impl(
-    state: &AppState,
-    id: &str,
-    db: &str,
-    collection: &str,
-) -> Result<Vec<IndexInfo>, String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        let key = format!("{}/{}/{}", id, db, collection);
-        let mut mock_indexes = state.mock_indexes.lock_safe()?;
-        if !mock_indexes.contains_key(&key) {
-            let defaults = mock_db::get_mock_indexes(db, collection);
-            mock_indexes.insert(key.clone(), defaults);
-        }
-        return Ok(mock_indexes.get(&key).unwrap().clone());
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let database = client.database(db);
-    let coll = database.collection::<mongodb::bson::Document>(collection);
-
-    let mut cursor = coll
-        .list_indexes()
-        .await
-        .map_err(|e| format!("Failed to list indexes: {}", e))?;
-
-    let mut indexes = Vec::new();
-    use futures::stream::StreamExt;
-    while let Some(result) = cursor.next().await {
-        let index_model = result.map_err(|e| format!("Index read error: {}", e))?;
-        // Serialize the real key pattern (preserves field order + direction/type).
-        let keys = serde_json::to_string(&index_model.keys).unwrap_or_else(|_| "{}".to_string());
-        let name = index_model
-            .options
-            .as_ref()
-            .and_then(|o| o.name.clone())
-            .unwrap_or_default();
-        let unique = index_model
-            .options
-            .as_ref()
-            .and_then(|o| o.unique)
-            .unwrap_or(false);
-        let sparse = index_model
-            .options
-            .as_ref()
-            .and_then(|o| o.sparse)
-            .unwrap_or(false);
-        indexes.push(IndexInfo {
-            name,
-            keys,
-            unique,
-            sparse,
-        });
-    }
-
-    Ok(indexes)
-}
-
-pub async fn execute_mql_query_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-    sort: &str,
-    projection: &str,
-    limit: i64,
-    skip: i64,
-) -> Result<Vec<String>, String> {
-    // Validate projection JSON up front (applies on the real path; mock ignores fields).
-    let projection_doc: Option<mongodb::bson::Document> =
-        if projection.trim().is_empty() || projection.trim() == "{}" {
-            None
-        } else {
-            let val: serde_json::Value = serde_json::from_str(projection)
-                .map_err(|e| format!("Invalid MQL projection JSON: {}", e))?;
-            Some(
-                mongodb::bson::to_document(&val)
-                    .map_err(|e| format!("BSON conversion error: {}", e))?,
-            )
-        };
-
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return mock_db::execute_mock_query(database, collection, filter, sort, limit, skip);
-    }
-
-    let filter_val: serde_json::Value = if filter.trim().is_empty() {
-        serde_json::Value::Object(serde_json::Map::new())
-    } else {
-        serde_json::from_str(filter).map_err(|e| format!("Invalid MQL filter JSON: {}", e))?
-    };
-
-    let sort_val: serde_json::Value = if sort.trim().is_empty() {
-        serde_json::Value::Object(serde_json::Map::new())
-    } else {
-        serde_json::from_str(sort).map_err(|e| format!("Invalid MQL sort JSON: {}", e))?
-    };
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let db = client.database(database);
-    let coll = db.collection::<mongodb::bson::Document>(collection);
-
-    // Convert filter serde_json::Value to BSON Document
-    let filter_doc = mongodb::bson::to_document(&filter_val)
-        .map_err(|e| format!("BSON conversion error: {}", e))?;
-
-    // Convert sort serde_json::Value to BSON Document
-    let sort_doc: Option<mongodb::bson::Document> =
-        if sort_val.is_object() && !sort_val.as_object().unwrap().is_empty() {
-            let doc = mongodb::bson::to_document(&sort_val)
-                .map_err(|e| format!("BSON conversion error: {}", e))?;
-            Some(doc)
-        } else {
-            None
-        };
-
-    let mut find_builder = coll.find(filter_doc);
-    if let Some(sort) = sort_doc {
-        find_builder = find_builder.sort(sort);
-    }
-    if let Some(projection) = projection_doc {
-        find_builder = find_builder.projection(projection);
-    }
-    if limit > 0 {
-        find_builder = find_builder.limit(limit);
-    }
-    if skip > 0 {
-        find_builder = find_builder.skip(skip as u64);
-    }
-
-    let mut cursor = find_builder
-        .await
-        .map_err(|e| format!("Query failed: {}", e))?;
-
-    let mut results = Vec::new();
-    use futures::stream::StreamExt;
-    while let Some(result) = cursor.next().await {
-        let doc = result.map_err(|e| format!("Cursor read error: {}", e))?;
-        let json_val: serde_json::Value =
-            serde_json::to_value(&doc).map_err(|e| format!("BSON to JSON error: {}", e))?;
-        results.push(serde_json::to_string(&json_val).unwrap());
-    }
-
-    Ok(results)
-}
-
-pub async fn count_documents_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-) -> Result<u64, String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return mock_db::count_mock_documents(database, collection, filter);
-    }
-
-    let trimmed = filter.trim();
-    let is_empty_filter = trimmed.is_empty() || trimmed == "{}";
-
-    let filter_val: serde_json::Value = if is_empty_filter {
-        serde_json::Value::Object(serde_json::Map::new())
-    } else {
-        serde_json::from_str(filter).map_err(|e| format!("Invalid MQL filter JSON: {}", e))?
-    };
-    let filter_doc = mongodb::bson::to_document(&filter_val)
-        .map_err(|e| format!("BSON conversion error: {}", e))?;
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let coll = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection);
-
-    // Empty filter: use the fast metadata estimate instead of a full scan.
-    let count = if is_empty_filter {
-        coll.estimated_document_count()
-            .await
-            .map_err(|e| format!("Failed to estimate document count: {}", e))?
-    } else {
-        coll.count_documents(filter_doc)
-            .await
-            .map_err(|e| format!("Count failed: {}", e))?
-    };
-    Ok(count)
-}
-
-fn csv_escape(raw: &str) -> String {
-    if raw.contains(',') || raw.contains('"') || raw.contains('\n') || raw.contains('\r') {
-        format!("\"{}\"", raw.replace('"', "\"\""))
-    } else {
-        raw.to_string()
-    }
-}
-
-fn csv_cell(value: Option<&serde_json::Value>) -> String {
-    match value {
-        None | Some(serde_json::Value::Null) => String::new(),
-        Some(serde_json::Value::String(s)) => csv_escape(s),
-        Some(serde_json::Value::Bool(v)) => csv_escape(&v.to_string()),
-        Some(serde_json::Value::Number(v)) => csv_escape(&v.to_string()),
-        Some(v) => csv_escape(&serde_json::to_string(v).unwrap_or_default()),
-    }
-}
-
-fn top_level_fields(value: &serde_json::Value) -> Vec<String> {
-    match value.as_object() {
-        Some(obj) => obj.keys().cloned().collect(),
-        None => Vec::new(),
-    }
-}
-
-fn bson_doc_to_json_value(doc: &mongodb::bson::Document) -> Result<serde_json::Value, String> {
-    serde_json::to_value(doc).map_err(|e| format!("BSON to JSON error: {}", e))
-}
-
-async fn export_mock_collection_to_file(
-    tasks: Arc<Mutex<HashMap<String, TaskInfo>>>,
-    task_id: String,
-    database: String,
-    collection: String,
-    format: String,
-    path: String,
-) -> Result<u64, String> {
-    let docs = mock_db::execute_mock_query(&database, &collection, "{}", "{}", i64::MAX, 0)?;
-    let total = docs.len() as u64;
-    update_task(&tasks, &task_id, |task| {
-        task.total = Some(total);
-        task.message = format!("Writing {} document(s)", total);
-    });
-
-    let mut file = tokio::fs::File::create(&path)
-        .await
-        .map_err(|e| format!("Failed to create export file: {}", e))?;
-
-    if format == "json" {
-        file.write_all(b"[\n")
-            .await
-            .map_err(|e| format!("Failed to write export file: {}", e))?;
-        for (idx, doc) in docs.iter().enumerate() {
-            if idx > 0 {
-                file.write_all(b",\n")
-                    .await
-                    .map_err(|e| format!("Failed to write export file: {}", e))?;
-            }
-            file.write_all(doc.as_bytes())
-                .await
-                .map_err(|e| format!("Failed to write export file: {}", e))?;
-            let processed = idx as u64 + 1;
-            update_task(&tasks, &task_id, |task| {
-                task.processed = processed;
-            });
-        }
-        file.write_all(b"\n]\n")
-            .await
-            .map_err(|e| format!("Failed to write export file: {}", e))?;
-    } else {
-        let values: Vec<serde_json::Value> = docs
-            .iter()
-            .map(|doc| {
-                serde_json::from_str(doc).map_err(|e| format!("Mock export JSON error: {}", e))
-            })
-            .collect::<Result<_, _>>()?;
-        let fields: Vec<String> = values
-            .iter()
-            .flat_map(top_level_fields)
-            .collect::<BTreeSet<_>>()
-            .into_iter()
-            .collect();
-        if !fields.is_empty() {
-            file.write_all(
-                fields
-                    .iter()
-                    .map(|field| csv_escape(field))
-                    .collect::<Vec<_>>()
-                    .join(",")
-                    .as_bytes(),
-            )
-            .await
-            .map_err(|e| format!("Failed to write export file: {}", e))?;
-            file.write_all(b"\n")
-                .await
-                .map_err(|e| format!("Failed to write export file: {}", e))?;
-        }
-        for (idx, value) in values.iter().enumerate() {
-            let obj = value.as_object();
-            let row = fields
-                .iter()
-                .map(|field| csv_cell(obj.and_then(|o| o.get(field))))
-                .collect::<Vec<_>>()
-                .join(",");
-            file.write_all(row.as_bytes())
-                .await
-                .map_err(|e| format!("Failed to write export file: {}", e))?;
-            file.write_all(b"\n")
-                .await
-                .map_err(|e| format!("Failed to write export file: {}", e))?;
-            let processed = idx as u64 + 1;
-            update_task(&tasks, &task_id, |task| {
-                task.processed = processed;
-            });
-        }
-    }
-
-    Ok(total)
-}
-
-async fn export_real_collection_to_file(
-    tasks: Arc<Mutex<HashMap<String, TaskInfo>>>,
-    task_id: String,
-    client: Client,
-    database: String,
-    collection: String,
-    format: String,
-    path: String,
-) -> Result<u64, String> {
-    use futures::stream::StreamExt;
-
-    let coll = client
-        .database(&database)
-        .collection::<mongodb::bson::Document>(&collection);
-
-    update_task(&tasks, &task_id, |task| {
-        task.message = "Counting documents".to_string();
-    });
-    let total = coll
-        .count_documents(mongodb::bson::Document::new())
-        .await
-        .map_err(|e| format!("Count failed: {}", e))?;
-    update_task(&tasks, &task_id, |task| {
-        task.total = Some(total);
-    });
-
-    let mut file = tokio::fs::File::create(&path)
-        .await
-        .map_err(|e| format!("Failed to create export file: {}", e))?;
-
-    if format == "json" {
-        update_task(&tasks, &task_id, |task| {
-            task.message = "Writing JSON".to_string();
-        });
-        file.write_all(b"[\n")
-            .await
-            .map_err(|e| format!("Failed to write export file: {}", e))?;
-        let mut cursor = coll
-            .find(mongodb::bson::Document::new())
-            .await
-            .map_err(|e| format!("Export query failed: {}", e))?;
-        let mut processed = 0u64;
-        while let Some(result) = cursor.next().await {
-            let doc = result.map_err(|e| format!("Cursor read error: {}", e))?;
-            let json_val = bson_doc_to_json_value(&doc)?;
-            let json = serde_json::to_string(&json_val)
-                .map_err(|e| format!("JSON serialization error: {}", e))?;
-            if processed > 0 {
-                file.write_all(b",\n")
-                    .await
-                    .map_err(|e| format!("Failed to write export file: {}", e))?;
-            }
-            file.write_all(json.as_bytes())
-                .await
-                .map_err(|e| format!("Failed to write export file: {}", e))?;
-            processed += 1;
-            if processed == total || processed % 100 == 0 {
-                update_task(&tasks, &task_id, |task| {
-                    task.processed = processed;
-                });
-            }
-        }
-        file.write_all(b"\n]\n")
-            .await
-            .map_err(|e| format!("Failed to write export file: {}", e))?;
-        update_task(&tasks, &task_id, |task| {
-            task.processed = processed;
-        });
-        Ok(processed)
-    } else {
-        update_task(&tasks, &task_id, |task| {
-            task.message = "Scanning CSV fields".to_string();
-        });
-        let mut fields = BTreeSet::new();
-        let mut scan_cursor = coll
-            .find(mongodb::bson::Document::new())
-            .await
-            .map_err(|e| format!("CSV field scan failed: {}", e))?;
-        let mut scanned = 0u64;
-        while let Some(result) = scan_cursor.next().await {
-            let doc = result.map_err(|e| format!("Cursor read error: {}", e))?;
-            let json_val = bson_doc_to_json_value(&doc)?;
-            for field in top_level_fields(&json_val) {
-                fields.insert(field);
-            }
-            scanned += 1;
-            if scanned == total || scanned % 250 == 0 {
-                update_task(&tasks, &task_id, |task| {
-                    task.processed = scanned;
-                });
-            }
-        }
-        let fields: Vec<String> = fields.into_iter().collect();
-        update_task(&tasks, &task_id, |task| {
-            task.processed = 0;
-            task.message = "Writing CSV".to_string();
-        });
-
-        if !fields.is_empty() {
-            file.write_all(
-                fields
-                    .iter()
-                    .map(|field| csv_escape(field))
-                    .collect::<Vec<_>>()
-                    .join(",")
-                    .as_bytes(),
-            )
-            .await
-            .map_err(|e| format!("Failed to write export file: {}", e))?;
-            file.write_all(b"\n")
-                .await
-                .map_err(|e| format!("Failed to write export file: {}", e))?;
-        }
-
-        let mut cursor = coll
-            .find(mongodb::bson::Document::new())
-            .await
-            .map_err(|e| format!("Export query failed: {}", e))?;
-        let mut processed = 0u64;
-        while let Some(result) = cursor.next().await {
-            let doc = result.map_err(|e| format!("Cursor read error: {}", e))?;
-            let json_val = bson_doc_to_json_value(&doc)?;
-            let obj = json_val.as_object();
-            let row = fields
-                .iter()
-                .map(|field| csv_cell(obj.and_then(|o| o.get(field))))
-                .collect::<Vec<_>>()
-                .join(",");
-            file.write_all(row.as_bytes())
-                .await
-                .map_err(|e| format!("Failed to write export file: {}", e))?;
-            file.write_all(b"\n")
-                .await
-                .map_err(|e| format!("Failed to write export file: {}", e))?;
-            processed += 1;
-            if processed == total || processed % 100 == 0 {
-                update_task(&tasks, &task_id, |task| {
-                    task.processed = processed;
-                });
-            }
-        }
-        update_task(&tasks, &task_id, |task| {
-            task.processed = processed;
-        });
-        Ok(processed)
-    }
-}
-
-pub async fn start_collection_export_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    format: &str,
-    path: &str,
-) -> Result<TaskInfo, String> {
-    let format = format.trim().to_lowercase();
-    if format != "json" && format != "csv" {
-        return Err("Export format must be json or csv".to_string());
-    }
-    if path.trim().is_empty() {
-        return Err("Export path is required".to_string());
-    }
-
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-    let client = if is_mock {
-        None
-    } else {
-        Some({
-            let connections = state.connections.lock_safe()?;
-            connections
-                .get(id)
-                .cloned()
-                .ok_or_else(|| "Connection client not found".to_string())?
-        })
-    };
-
-    let task_id = Uuid::new_v4().to_string();
-    let task = TaskInfo {
-        id: task_id.clone(),
-        kind: "collection_export".to_string(),
-        label: format!(
-            "Export {}.{} as {}",
-            database,
-            collection,
-            format.to_uppercase()
-        ),
-        status: "running".to_string(),
-        processed: 0,
-        total: None,
-        message: "Queued".to_string(),
-        path: Some(path.to_string()),
-        error: None,
-        created_at_ms: now_ms(),
-        finished_at_ms: None,
-    };
-    state
-        .tasks
-        .lock()
-        .unwrap()
-        .insert(task_id.clone(), task.clone());
-
-    let tasks = state.tasks.clone();
-    let database = database.to_string();
-    let collection = collection.to_string();
-    let path = path.to_string();
-    tokio::spawn(async move {
-        let result = if let Some(client) = client {
-            export_real_collection_to_file(
-                tasks.clone(),
-                task_id.clone(),
-                client,
-                database,
-                collection,
-                format,
-                path,
-            )
-            .await
-        } else {
-            export_mock_collection_to_file(
-                tasks.clone(),
-                task_id.clone(),
-                database,
-                collection,
-                format,
-                path,
-            )
-            .await
-        };
-        match result {
-            Ok(processed) => finish_task(&tasks, &task_id, processed),
-            Err(err) => fail_task(&tasks, &task_id, err),
-        }
-    });
-
-    Ok(task)
-}
-
-pub async fn execute_aggregate_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    pipeline: &str,
-) -> Result<Vec<String>, String> {
-    // Parse and validate the pipeline (a JSON array of stage documents) up front so a
-    // malformed pipeline fails clearly regardless of the connection type.
-    let pipeline_val: serde_json::Value = if pipeline.trim().is_empty() {
-        serde_json::Value::Array(Vec::new())
-    } else {
-        serde_json::from_str(pipeline)
-            .map_err(|e| format!("Invalid aggregation pipeline JSON: {}", e))?
-    };
-    let stages_val = pipeline_val
-        .as_array()
-        .ok_or_else(|| "Aggregation pipeline must be a JSON array of stages".to_string())?;
-    let mut stages: Vec<mongodb::bson::Document> = Vec::with_capacity(stages_val.len());
-    for stage in stages_val {
-        let doc = mongodb::bson::to_document(stage)
-            .map_err(|e| format!("Invalid aggregation stage: {}", e))?;
-        stages.push(doc);
-    }
-
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return Err("Aggregation pipelines are not supported on mock connections".to_string());
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let coll = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection);
-
-    let mut cursor = coll
-        .aggregate(stages)
-        .await
-        .map_err(|e| format!("Aggregation failed: {}", e))?;
-
-    let mut results = Vec::new();
-    use futures::stream::StreamExt;
-    while let Some(result) = cursor.next().await {
-        let doc = result.map_err(|e| format!("Cursor read error: {}", e))?;
-        let json_val: serde_json::Value =
-            serde_json::to_value(&doc).map_err(|e| format!("BSON to JSON error: {}", e))?;
-        results.push(serde_json::to_string(&json_val).unwrap());
-    }
-
-    Ok(results)
-}
-
-/// Explain an entire aggregation pipeline (M1), not just its `$match` stage.
-/// Mirrors `execute_aggregate_impl`'s validation and is real-connection-only.
-pub async fn explain_aggregate_query_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    pipeline: &str,
-) -> Result<String, String> {
-    // Parse and validate the pipeline (a JSON array of stage documents) up front.
-    let pipeline_val: serde_json::Value = if pipeline.trim().is_empty() {
-        serde_json::Value::Array(Vec::new())
-    } else {
-        serde_json::from_str(pipeline)
-            .map_err(|e| format!("Invalid aggregation pipeline JSON: {}", e))?
-    };
-    let stages_val = pipeline_val
-        .as_array()
-        .ok_or_else(|| "Aggregation pipeline must be a JSON array of stages".to_string())?;
-    let mut stages: Vec<mongodb::bson::Document> = Vec::with_capacity(stages_val.len());
-    for stage in stages_val {
-        let doc = mongodb::bson::to_document(stage)
-            .map_err(|e| format!("Invalid aggregation stage: {}", e))?;
-        stages.push(doc);
-    }
-
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return Err("Aggregation explain is not supported on mock connections".to_string());
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let db = client.database(database);
-    let command = mongodb::bson::doc! {
-        "explain": {
-            "aggregate": collection,
-            "pipeline": stages,
-            "cursor": {}
-        },
-        "verbosity": "executionStats"
-    };
-
-    let explain_result = db
-        .run_command(command)
-        .await
-        .map_err(|e| format!("Explain failed: {}", e))?;
-
-    let json_val: serde_json::Value =
-        serde_json::to_value(&explain_result).map_err(|e| format!("BSON to JSON error: {}", e))?;
-
-    Ok(serde_json::to_string_pretty(&json_val).unwrap())
-}
-
-pub async fn explain_mql_query_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-) -> Result<String, String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        return Ok(mock_db::get_mock_explain(database, collection, filter));
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let db = client.database(database);
-    let filter_val: serde_json::Value = if filter.trim().is_empty() {
-        serde_json::Value::Object(serde_json::Map::new())
-    } else {
-        serde_json::from_str(filter).map_err(|e| format!("Invalid MQL filter JSON: {}", e))?
-    };
-
-    let filter_doc = mongodb::bson::to_document(&filter_val)
-        .map_err(|e| format!("BSON conversion error: {}", e))?;
-
-    let command = mongodb::bson::doc! {
-        "explain": {
-            "find": collection,
-            "filter": filter_doc
-        },
-        "verbosity": "executionStats"
-    };
-
-    let explain_result = db
-        .run_command(command)
-        .await
-        .map_err(|e| format!("Explain failed: {}", e))?;
-
-    let json_val: serde_json::Value =
-        serde_json::to_value(&explain_result).map_err(|e| format!("BSON to JSON error: {}", e))?;
-
-    Ok(serde_json::to_string_pretty(&json_val).unwrap())
-}
-
-pub async fn create_index_impl(
-    state: &AppState,
-    id: &str,
-    db: &str,
-    collection: &str,
-    index_name: &str,
-    keys: &str,
-    unique: bool,
-    sparse: bool,
-) -> Result<(), String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        let key = format!("{}/{}/{}", id, db, collection);
-        let mut mock_indexes = state.mock_indexes.lock_safe()?;
-        if !mock_indexes.contains_key(&key) {
-            let defaults = mock_db::get_mock_indexes(db, collection);
-            mock_indexes.insert(key.clone(), defaults);
-        }
-        let list = mock_indexes.get_mut(&key).unwrap();
-        if !list.iter().any(|i| i.name == index_name) {
-            list.push(IndexInfo {
-                name: index_name.to_string(),
-                keys: keys.to_string(),
-                unique,
-                sparse,
-            });
-        }
-        return Ok(());
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let database = client.database(db);
-    let coll = database.collection::<mongodb::bson::Document>(collection);
-
-    let value: serde_json::Value =
-        serde_json::from_str(keys).map_err(|e| format!("Invalid JSON keys: {}", e))?;
-    let keys_doc = mongodb::bson::to_document(&value)
-        .map_err(|e| format!("Failed to convert keys JSON to BSON: {}", e))?;
-
-    let mut options = mongodb::options::IndexOptions::builder()
-        .name(index_name.to_string())
-        .build();
-
-    if unique {
-        options.unique = Some(true);
-    }
-    if sparse {
-        options.sparse = Some(true);
-    }
-
-    let model = mongodb::IndexModel::builder()
-        .keys(keys_doc)
-        .options(options)
-        .build();
-
-    coll.create_index(model)
-        .await
-        .map(|_| ())
-        .map_err(|e| format!("Failed to create index: {}", e))
-}
-
-pub async fn delete_index_impl(
-    state: &AppState,
-    id: &str,
-    db: &str,
-    collection: &str,
-    index_name: &str,
-) -> Result<(), String> {
-    let is_mock = {
-        let mocks = state.mocks.lock_safe()?;
-        *mocks
-            .get(id)
-            .ok_or_else(|| "Connection not found".to_string())?
-    };
-
-    if is_mock {
-        let key = format!("{}/{}/{}", id, db, collection);
-        let mut mock_indexes = state.mock_indexes.lock_safe()?;
-        if !mock_indexes.contains_key(&key) {
-            let defaults = mock_db::get_mock_indexes(db, collection);
-            mock_indexes.insert(key.clone(), defaults);
-        }
-        let list = mock_indexes.get_mut(&key).unwrap();
-        list.retain(|x| x.name != index_name);
-        return Ok(());
-    }
-
-    let client = {
-        let connections = state.connections.lock_safe()?;
-        connections
-            .get(id)
-            .cloned()
-            .ok_or_else(|| "Connection client not found".to_string())?
-    };
-
-    let database = client.database(db);
-    let coll = database.collection::<mongodb::bson::Document>(collection);
-
-    coll.drop_index(index_name)
-        .await
-        .map(|_| ())
-        .map_err(|e| format!("Failed to delete index: {}", e))
-}
-
-// Parse a JSON string into a BSON Document, interpreting MongoDB Extended JSON
-// (e.g. {"$oid": "..."} -> ObjectId, {"$date": ...} -> DateTime) so that writes
-// match documents by their real _id type rather than a literal sub-document.
-pub fn json_to_bson_document(s: &str) -> Result<mongodb::bson::Document, String> {
-    let value: serde_json::Value =
-        serde_json::from_str(s).map_err(|e| format!("Invalid JSON: {}", e))?;
-    let bson = mongodb::bson::Bson::try_from(value)
-        .map_err(|e| format!("Invalid BSON/Extended JSON: {}", e))?;
-    match bson {
-        mongodb::bson::Bson::Document(doc) => Ok(doc),
-        _ => Err("Expected a JSON object (e.g. { \"field\": value })".to_string()),
-    }
-}
-
 pub(crate) fn require_real_client(state: &AppState, id: &str) -> Result<Client, String> {
     let connections = state.connections.lock_safe()?;
     connections
@@ -1538,587 +521,6 @@ pub(crate) fn connection_is_mock(state: &AppState, id: &str) -> Result<bool, Str
         .get(id)
         .copied()
         .ok_or_else(|| "Connection not found".to_string())
-}
-
-pub async fn delete_document_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-) -> Result<u64, String> {
-    // Parse/validate up front so bad input fails the same way for mock & real.
-    let filter_doc = json_to_bson_document(filter)?;
-
-    if connection_is_mock(state, id)? {
-        return Ok(1);
-    }
-
-    let client = require_real_client(state, id)?;
-    let coll = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection);
-    let res = coll
-        .delete_one(filter_doc)
-        .await
-        .map_err(|e| format!("Failed to delete document: {}", e))?;
-    Ok(res.deleted_count)
-}
-
-pub async fn delete_many_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-) -> Result<u64, String> {
-    let filter_doc = json_to_bson_document(filter)?;
-    if connection_is_mock(state, id)? {
-        return Ok(0); // mock connections don't persist deletes
-    }
-    let client = require_real_client(state, id)?;
-    let res = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection)
-        .delete_many(filter_doc)
-        .await
-        .map_err(|e| format!("Failed to delete documents: {}", e))?;
-    Ok(res.deleted_count)
-}
-
-pub async fn update_many_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-    update: &str,
-) -> Result<u64, String> {
-    let filter_doc = json_to_bson_document(filter)?;
-    let update_doc = json_to_bson_document(update)?;
-    // Require an operator-keyed update ({ "$set": … }); reject bare replacements
-    // / empty updates so a bulk op can't silently overwrite whole documents.
-    if update_doc.is_empty() || !update_doc.keys().all(|k| k.starts_with('$')) {
-        return Err("Update must use operators like $set (e.g. { \"$set\": { … } })".to_string());
-    }
-    if connection_is_mock(state, id)? {
-        return Ok(0); // mock connections don't persist updates
-    }
-    let client = require_real_client(state, id)?;
-    let res = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection)
-        .update_many(filter_doc, update_doc)
-        .await
-        .map_err(|e| format!("Failed to update documents: {}", e))?;
-    Ok(res.modified_count)
-}
-
-pub async fn insert_document_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    document: &str,
-) -> Result<String, String> {
-    let doc = json_to_bson_document(document)?;
-
-    if connection_is_mock(state, id)? {
-        return Ok("mock-inserted-id".to_string());
-    }
-
-    let client = require_real_client(state, id)?;
-    let coll = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection);
-    let res = coll
-        .insert_one(doc)
-        .await
-        .map_err(|e| format!("Failed to insert document: {}", e))?;
-    // Return the inserted id as a JSON string (Extended JSON for ObjectId etc.).
-    Ok(res.inserted_id.into_relaxed_extjson().to_string())
-}
-
-pub async fn update_document_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    filter: &str,
-    replacement: &str,
-) -> Result<u64, String> {
-    let filter_doc = json_to_bson_document(filter)?;
-    let replacement_doc = json_to_bson_document(replacement)?;
-
-    if connection_is_mock(state, id)? {
-        return Ok(1);
-    }
-
-    let client = require_real_client(state, id)?;
-    let coll = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection);
-    let res = coll
-        .replace_one(filter_doc, replacement_doc)
-        .await
-        .map_err(|e| format!("Failed to update document: {}", e))?;
-    Ok(res.modified_count)
-}
-
-#[derive(serde::Serialize)]
-pub struct ImportResult {
-    pub inserted: u64,
-    pub updated: u64,
-    pub skipped: u64,
-}
-
-#[derive(serde::Serialize)]
-pub struct DatabaseRenameResult {
-    pub collections: u64,
-    pub documents: u64,
-}
-
-// Bulk-import documents with a duplicate-handling mode:
-//   "skip"   - insert_many(ordered:false); duplicate-key rows are counted as skipped.
-//   "update" - per doc replace_one({_id}, doc, upsert:true); no _id -> insert.
-//   "abort"  - if any incoming _id already exists, write nothing and error.
-// Documents are already-validated JSON values from the frontend codec; each is
-// converted to BSON here as a safety net.
-pub async fn import_documents_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-    docs: Vec<serde_json::Value>,
-    mode: &str,
-) -> Result<ImportResult, String> {
-    // Convert all docs up front; a bad doc fails the whole import before writing.
-    let mut bson_docs: Vec<mongodb::bson::Document> = Vec::with_capacity(docs.len());
-    for value in &docs {
-        let s = serde_json::to_string(value).map_err(|e| format!("Invalid document: {}", e))?;
-        bson_docs.push(json_to_bson_document(&s)?);
-    }
-
-    if connection_is_mock(state, id)? {
-        // Mock connections validate but do not persist.
-        return Ok(match mode {
-            "update" => ImportResult {
-                inserted: 0,
-                updated: bson_docs.len() as u64,
-                skipped: 0,
-            },
-            _ => ImportResult {
-                inserted: bson_docs.len() as u64,
-                updated: 0,
-                skipped: 0,
-            },
-        });
-    }
-
-    let client = require_real_client(state, id)?;
-    let coll = client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection);
-
-    // Collect the set of incoming _ids that already exist in the collection,
-    // keyed by their stringified BSON (same rendering on both sides). Used by
-    // skip (partition) and abort (pre-check) so we never rely on bulk-write
-    // error introspection, which varies across driver versions.
-    async fn existing_ids(
-        coll: &mongodb::Collection<mongodb::bson::Document>,
-        docs: &[mongodb::bson::Document],
-    ) -> Result<std::collections::HashSet<String>, String> {
-        let ids: Vec<mongodb::bson::Bson> =
-            docs.iter().filter_map(|d| d.get("_id").cloned()).collect();
-        let mut found = std::collections::HashSet::new();
-        if ids.is_empty() {
-            return Ok(found);
-        }
-        let mut cursor = coll
-            .find(mongodb::bson::doc! { "_id": { "$in": ids } })
-            .await
-            .map_err(|e| format!("Failed to check existing documents: {}", e))?;
-        use futures::stream::StreamExt;
-        while let Some(result) = cursor.next().await {
-            let doc = result.map_err(|e| format!("Cursor read error: {}", e))?;
-            if let Some(id_val) = doc.get("_id") {
-                found.insert(id_val.to_string());
-            }
-        }
-        Ok(found)
-    }
-
-    match mode {
-        "update" => {
-            // Existing _ids get replaced (counts as updated); everything else is
-            // inserted. This is an upsert-by-_id without relying on the driver's
-            // option setters, which vary across versions.
-            let existing = existing_ids(&coll, &bson_docs).await?;
-            let mut inserted = 0u64;
-            let mut updated = 0u64;
-            for doc in bson_docs {
-                match doc.get("_id").cloned() {
-                    Some(id_val) if existing.contains(&id_val.to_string()) => {
-                        let filter = mongodb::bson::doc! { "_id": id_val };
-                        let res = coll
-                            .replace_one(filter, doc)
-                            .await
-                            .map_err(|e| format!("Failed to import (update): {}", e))?;
-                        updated += res.modified_count;
-                    }
-                    _ => {
-                        coll.insert_one(doc)
-                            .await
-                            .map_err(|e| format!("Failed to import (insert): {}", e))?;
-                        inserted += 1;
-                    }
-                }
-            }
-            Ok(ImportResult {
-                inserted,
-                updated,
-                skipped: 0,
-            })
-        }
-        "abort" => {
-            // Any incoming _id already present -> abort, write nothing.
-            let existing = existing_ids(&coll, &bson_docs).await?;
-            if !existing.is_empty() {
-                return Err(format!(
-                    "Import aborted: {} document(s) already exist",
-                    existing.len()
-                ));
-            }
-            let total = bson_docs.len() as u64;
-            coll.insert_many(bson_docs)
-                .await
-                .map_err(|e| format!("Failed to import: {}", e))?;
-            Ok(ImportResult {
-                inserted: total,
-                updated: 0,
-                skipped: 0,
-            })
-        }
-        _ => {
-            // "skip" (default): insert only docs whose _id does not already exist;
-            // count existing ones as skipped. Docs without an _id always insert.
-            let existing = existing_ids(&coll, &bson_docs).await?;
-            let total = bson_docs.len() as u64;
-            let to_insert: Vec<mongodb::bson::Document> = bson_docs
-                .into_iter()
-                .filter(|d| match d.get("_id") {
-                    Some(id_val) => !existing.contains(&id_val.to_string()),
-                    None => true,
-                })
-                .collect();
-            let inserted = to_insert.len() as u64;
-            if !to_insert.is_empty() {
-                coll.insert_many(to_insert)
-                    .await
-                    .map_err(|e| format!("Failed to import: {}", e))?;
-            }
-            Ok(ImportResult {
-                inserted,
-                updated: 0,
-                skipped: total - inserted,
-            })
-        }
-    }
-}
-
-pub async fn create_collection_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-) -> Result<(), String> {
-    if collection.trim().is_empty() {
-        return Err("Collection name is required".to_string());
-    }
-    if connection_is_mock(state, id)? {
-        return Ok(());
-    }
-    let client = require_real_client(state, id)?;
-    client
-        .database(database)
-        .create_collection(collection)
-        .await
-        .map_err(|e| format!("Failed to create collection: {}", e))
-}
-
-pub async fn create_view_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    view_name: &str,
-    source_collection: &str,
-    pipeline: &str,
-) -> Result<(), String> {
-    if view_name.trim().is_empty() {
-        return Err("View name is required".to_string());
-    }
-    if source_collection.trim().is_empty() {
-        return Err("Source collection is required".to_string());
-    }
-    // Validate the pipeline (a JSON array of stage documents) up front.
-    let pipeline_val: serde_json::Value = if pipeline.trim().is_empty() {
-        serde_json::Value::Array(Vec::new())
-    } else {
-        serde_json::from_str(pipeline)
-            .map_err(|e| format!("Invalid aggregation pipeline JSON: {}", e))?
-    };
-    let stages_val = pipeline_val
-        .as_array()
-        .ok_or_else(|| "Aggregation pipeline must be a JSON array of stages".to_string())?;
-    let mut stages: Vec<mongodb::bson::Document> = Vec::with_capacity(stages_val.len());
-    for stage in stages_val {
-        let doc = mongodb::bson::to_document(stage)
-            .map_err(|e| format!("Invalid aggregation stage: {}", e))?;
-        stages.push(doc);
-    }
-
-    if connection_is_mock(state, id)? {
-        return Ok(());
-    }
-    let client = require_real_client(state, id)?;
-    client
-        .database(database)
-        .create_collection(view_name)
-        .view_on(source_collection.to_string())
-        .pipeline(stages)
-        .await
-        .map_err(|e| format!("Failed to create view: {}", e))
-}
-
-pub async fn drop_collection_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    collection: &str,
-) -> Result<(), String> {
-    if connection_is_mock(state, id)? {
-        return Ok(());
-    }
-    let client = require_real_client(state, id)?;
-    client
-        .database(database)
-        .collection::<mongodb::bson::Document>(collection)
-        .drop()
-        .await
-        .map_err(|e| format!("Failed to drop collection: {}", e))
-}
-
-pub async fn rename_collection_impl(
-    state: &AppState,
-    id: &str,
-    database: &str,
-    from: &str,
-    to: &str,
-) -> Result<(), String> {
-    if from.trim().is_empty() || to.trim().is_empty() {
-        return Err("Collection name is required".to_string());
-    }
-    if from == to {
-        return Err("New collection name must be different".to_string());
-    }
-    if connection_is_mock(state, id)? {
-        return Ok(());
-    }
-    let client = require_real_client(state, id)?;
-    client
-        .database("admin")
-        .run_command(mongodb::bson::doc! {
-            "renameCollection": format!("{}.{}", database, from),
-            "to": format!("{}.{}", database, to),
-            "dropTarget": false,
-        })
-        .await
-        .map(|_| ())
-        .map_err(|e| format!("Failed to rename collection: {}", e))
-}
-
-pub async fn drop_database_impl(state: &AppState, id: &str, database: &str) -> Result<(), String> {
-    if database.trim().is_empty() {
-        return Err("Database name is required".to_string());
-    }
-    if connection_is_mock(state, id)? {
-        return Ok(());
-    }
-    let client = require_real_client(state, id)?;
-    client
-        .database(database)
-        .drop()
-        .await
-        .map_err(|e| format!("Failed to drop database: {}", e))
-}
-
-pub async fn rename_database_impl(
-    state: &AppState,
-    id: &str,
-    from: &str,
-    to: &str,
-    drop_source: bool,
-) -> Result<DatabaseRenameResult, String> {
-    if from.trim().is_empty() || to.trim().is_empty() {
-        return Err("Database name is required".to_string());
-    }
-    if from == to {
-        return Err("New database name must be different".to_string());
-    }
-    if connection_is_mock(state, id)? {
-        return Ok(DatabaseRenameResult {
-            collections: 0,
-            documents: 0,
-        });
-    }
-
-    let client = require_real_client(state, id)?;
-    let db_names = client
-        .list_database_names()
-        .await
-        .map_err(|e| format!("Failed to list databases: {}", e))?;
-    if !db_names.iter().any(|name| name == from) {
-        return Err(format!("Source database \"{}\" does not exist", from));
-    }
-    if db_names.iter().any(|name| name == to) {
-        return Err(format!("Target database \"{}\" already exists", to));
-    }
-
-    let source_db = client.database(from);
-    let target_db = client.database(to);
-    let mut coll_cursor = source_db
-        .list_collections()
-        .await
-        .map_err(|e| format!("Failed to list source collections: {}", e))?;
-
-    let mut specs = Vec::new();
-    use futures::stream::StreamExt;
-    while let Some(result) = coll_cursor.next().await {
-        let spec = result.map_err(|e| format!("Collection read error: {}", e))?;
-        match spec.collection_type {
-            mongodb::results::CollectionType::Collection => specs.push(spec),
-            mongodb::results::CollectionType::View => {
-                return Err(format!(
-                    "Cannot rename database: collection \"{}\" is a view",
-                    spec.name
-                ));
-            }
-            mongodb::results::CollectionType::Timeseries => {
-                return Err(format!(
-                    "Cannot rename database: collection \"{}\" is time-series",
-                    spec.name
-                ));
-            }
-            _ => {
-                return Err(format!(
-                    "Cannot rename database: collection \"{}\" has an unsupported type",
-                    spec.name
-                ));
-            }
-        }
-    }
-    if specs.is_empty() {
-        return Err(format!(
-            "Source database \"{}\" has no collections to copy",
-            from
-        ));
-    }
-
-    let copy_result = async {
-        let mut copied_collections = 0u64;
-        let mut copied_documents = 0u64;
-
-        for spec in &specs {
-            let source_coll = source_db.collection::<mongodb::bson::Document>(&spec.name);
-            let target_coll = target_db.collection::<mongodb::bson::Document>(&spec.name);
-
-            target_db
-                .create_collection(&spec.name)
-                .await
-                .map_err(|e| format!("Failed to create target collection {}: {}", spec.name, e))?;
-
-            let mut doc_cursor = source_coll
-                .find(mongodb::bson::doc! {})
-                .await
-                .map_err(|e| format!("Failed to read source collection {}: {}", spec.name, e))?;
-            let mut batch = Vec::with_capacity(500);
-            while let Some(result) = doc_cursor.next().await {
-                batch.push(result.map_err(|e| format!("Cursor read error: {}", e))?);
-                if batch.len() >= 500 {
-                    copied_documents += batch.len() as u64;
-                    target_coll
-                        .insert_many(std::mem::take(&mut batch))
-                        .await
-                        .map_err(|e| {
-                            format!("Failed to copy documents for {}: {}", spec.name, e)
-                        })?;
-                }
-            }
-            if !batch.is_empty() {
-                copied_documents += batch.len() as u64;
-                target_coll
-                    .insert_many(batch)
-                    .await
-                    .map_err(|e| format!("Failed to copy documents for {}: {}", spec.name, e))?;
-            }
-
-            let mut idx_cursor = source_coll
-                .list_indexes()
-                .await
-                .map_err(|e| format!("Failed to list indexes for {}: {}", spec.name, e))?;
-            while let Some(result) = idx_cursor.next().await {
-                let index = result.map_err(|e| format!("Index read error: {}", e))?;
-                let name = index
-                    .options
-                    .as_ref()
-                    .and_then(|o| o.name.as_deref())
-                    .unwrap_or("");
-                if name != "_id_" {
-                    target_coll
-                        .create_index(index)
-                        .await
-                        .map_err(|e| format!("Failed to recreate index on {}: {}", spec.name, e))?;
-                }
-            }
-
-            let source_count = source_coll
-                .count_documents(mongodb::bson::doc! {})
-                .await
-                .map_err(|e| format!("Failed to count source collection {}: {}", spec.name, e))?;
-            let target_count = target_coll
-                .count_documents(mongodb::bson::doc! {})
-                .await
-                .map_err(|e| format!("Failed to count target collection {}: {}", spec.name, e))?;
-            if source_count != target_count {
-                return Err(format!(
-                    "Copied collection {} failed verification: source has {}, target has {}",
-                    spec.name, source_count, target_count
-                ));
-            }
-            copied_collections += 1;
-        }
-
-        Ok(DatabaseRenameResult {
-            collections: copied_collections,
-            documents: copied_documents,
-        })
-    }
-    .await;
-
-    match copy_result {
-        Ok(result) => {
-            if drop_source {
-                source_db.drop().await.map_err(|e| {
-                    format!("Copied target, but failed to drop source database: {}", e)
-                })?;
-            }
-            Ok(result)
-        }
-        Err(err) => {
-            let _ = target_db.drop().await;
-            Err(err)
-        }
-    }
 }
 
 // Tauri Command wrappers (kept private to module to avoid reimport collisions)
@@ -2378,7 +780,15 @@ async fn create_view(
     source_collection: String,
     pipeline: String,
 ) -> Result<(), String> {
-    create_view_impl(&state, &id, &database, &view_name, &source_collection, &pipeline).await
+    create_view_impl(
+        &state,
+        &id,
+        &database,
+        &view_name,
+        &source_collection,
+        &pipeline,
+    )
+    .await
 }
 
 #[tauri::command]
